+++ pkg/contrast/contrast_test.go
@@ -2,7 +2,10 @@ package contrast
 
 import (
 	"fmt"
+	"net/http"
+	"net/http/httptest"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -346,5 +349,201 @@ func TestAccumulateFindings(t *testing.T) {
 		assert.Equal(t, 55, findings[0].Audited)
 		assert.Equal(t, 200, findings[1].Total)
 		assert.Equal(t, 100, findings[1].Audited)
+	})
+}
+
+// Client tests
+
+func TestClientCreation(t *testing.T) {
+	t.Parallel()
+	t.Run("with_custom_base_URL", func(t *testing.T) {
+		customURL := "https://custom.contrastsecurity.com"
+		appURL := "https://api.example.com/app"
+		client := NewClient("api-key", "service-key", "user@example.com", "org-123", customURL, appURL)
+
+		assert.NotNil(t, client)
+		assert.Equal(t, "api-key", client.ApiKey)
+		assert.Equal(t, "service-key", client.ServiceKey)
+		assert.Equal(t, "user@example.com", client.Username)
+		assert.Equal(t, "org-123", client.OrgID)
+		assert.Equal(t, customURL, client.BaseURL)
+		assert.Equal(t, appURL, client.AppURL)
+		assert.NotEmpty(t, client.Auth, "Auth should be populated with base64 encoded credentials")
+		assert.NotNil(t, client.HttpClient)
+	})
+
+	t.Run("with_empty_base_URL_uses_default", func(t *testing.T) {
+		appURL := "https://api.example.com/app"
+		client := NewClient("api-key", "service-key", "user@example.com", "org-123", "", appURL)
+
+		assert.NotNil(t, client)
+		assert.Equal(t, "https://cs003.contrastsecurity.com", client.BaseURL)
+	})
+}
+
+func TestAddAuth(t *testing.T) {
+	t.Parallel()
+	client := NewClient("api-key", "service-key", "user@example.com", "org-123", "", "")
+	req, _ := http.NewRequest("GET", "https://example.com", nil)
+
+	client.addAuth(req)
+
+	assert.NotEmpty(t, req.Header.Get("Authorization"))
+	assert.Equal(t, "Basic "+client.Auth, req.Header.Get("Authorization"))
+	assert.Equal(t, "api-key", req.Header.Get("API-Key"))
+}
+
+func TestCheckReportStatusSuccess(t *testing.T) {
+	t.Parallel()
+	downloadURL := ""
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprintf(w, `{"success": true, "status": "ACTIVE", "downloadUrl": "%s/download"}`, downloadURL)
+	}))
+	defer server.Close()
+	downloadURL = server.URL
+
+	client := NewClient("api-key", "service-key", "user@example.com", "org-123", server.URL, "")
+
+	resp, err := client.checkReportStatus(server.URL + "/status")
+
+	assert.NoError(t, err)
+	assert.NotNil(t, resp)
+	assert.True(t, resp.Success)
+	assert.Equal(t, "ACTIVE", resp.Status)
+	assert.NotEmpty(t, resp.DownloadUrl)
+}
+
+func TestCheckReportStatusServerError(t *testing.T) {
+	t.Parallel()
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusInternalServerError)
+	}))
+	defer server.Close()
+
+	client := NewClient("api-key", "service-key", "user@example.com", "org-123", server.URL, "")
+
+	_, err := client.checkReportStatus(server.URL + "/status")
+
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "unexpected status code: 500")
+}
+
+func TestCheckReportStatusInvalidJSON(t *testing.T) {
+	t.Parallel()
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprintf(w, `invalid json`)
+	}))
+	defer server.Close()
+
+	client := NewClient("api-key", "service-key", "user@example.com", "org-123", server.URL, "")
+
+	_, err := client.checkReportStatus(server.URL + "/status")
+
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "failed to parse response")
+}
+
+func TestPollReportStatusSuccess(t *testing.T) {
+	t.Parallel()
+	downloadURL := ""
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprintf(w, `{"success": true, "status": "ACTIVE", "downloadUrl": "%s/download"}`, downloadURL)
+	}))
+	defer server.Close()
+	downloadURL = server.URL
+
+	client := NewClient("api-key", "service-key", "user@example.com", "org-123", server.URL, "")
+
+	resp, err := client.PollReportStatus("test-uuid", "TEST")
+
+	assert.NoError(t, err)
+	assert.NotNil(t, resp)
+	assert.Equal(t, "ACTIVE", resp.Status)
+}
+
+func TestDownloadReportSuccess(t *testing.T) {
+	t.Parallel()
+	expectedData := []byte("test report content")
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusOK)
+		w.Write(expectedData)
+	}))
+	defer server.Close()
+
+	client := NewClient("api-key", "service-key", "user@example.com", "org-123", server.URL, "")
+
+	data, err := client.DownloadReport(server.URL+"/download", "TEST")
+
+	assert.NoError(t, err)
+	assert.Equal(t, expectedData, data)
+}
+
+func TestDownloadReportServerError(t *testing.T) {
+	t.Parallel()
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprintf(w, `{"error": "server error"}`)
+	}))
+	defer server.Close()
+
+	client := NewClient("api-key", "service-key", "user@example.com", "org-123", server.URL, "")
+
+	_, err := client.DownloadReport(server.URL+"/download", "TEST")
+
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "unexpected status code: 500")
+}
+
+func TestNewPollConfig(t *testing.T) {
+	t.Parallel()
+	config := newPollConfig()
+
+	assert.Equal(t, 5*time.Minute, config.maxTotalWait)
+	assert.Equal(t, 60*time.Second, config.maxPollInterval)
+	assert.Equal(t, 15*time.Second, config.initialDelay)
+	assert.Equal(t, 5*time.Second, config.pollInterval)
+	assert.Equal(t, 1.5, config.backoffFactor)
+}
+
+func testMaxIntervalCapping(t *testing.T, config pollConfig) {
+	// Test interval capping logic without actual sleep
+	// This tests the backoff calculation, not the sleep behavior
+	pollInterval := config.pollInterval
+
+	// Simulate multiple backoff iterations
+	for i := 0; i < 20; i++ {
+		nextInterval := time.Duration(float64(pollInterval) * config.backoffFactor)
+		if nextInterval > config.maxPollInterval {
+			pollInterval = config.maxPollInterval
+		} else {
+			pollInterval = nextInterval
+		}
+
+		if pollInterval > config.maxPollInterval {
+			t.Errorf("Poll interval exceeded max at iteration %d: %v > %v", i, pollInterval, config.maxPollInterval)
+		}
+		assert.LessOrEqual(t, pollInterval, config.maxPollInterval)
+	}
+}
+
+func TestWaitAndBackoff(t *testing.T) {
+	t.Parallel()
+	config := newPollConfig()
+
+	t.Run("backoff_factor_applies", func(t *testing.T) {
+		initialInterval := config.pollInterval
+		// Test interval calculation logic without sleeping
+		nextInterval := time.Duration(float64(initialInterval) * config.backoffFactor)
+
+		assert.Greater(t, nextInterval, initialInterval, "interval should increase due to backoff factor")
+	})
+
+	t.Run("max_interval_capping", func(t *testing.T) {
+		testMaxIntervalCapping(t, config)
 	})
 }
